入力待ち状態をリサイズ再描画で保持
