+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- [ 不具合修正 ] ペインをリサイズすると「🟡 入力待ち」インジケータが消える不具合を修正。入力待ち状態を入力まで保持し、リサイズ起因の再描画（確認文の折り返し変化）による誤解除を防止（[#91](https://github.com/vektor-inc/vk-terminals/issues/91)）
+
 ## 1.11.0
 
 - [ 機能追加 ] グリッドのペインをサイドバーに格納して表示エリアを空け、必要なときにグリッドへ戻せる機能を追加。格納中もターミナルは稼働を継続し、サイドバー幅のドラッグ変更にも対応（[#89](https://github.com/vektor-inc/vk-terminals/issues/89)）
