+++ renderer/waitingState.js
@@ -1,0 +1,72 @@
+/* global module */
+
+// ─── Waiting detection ────────────────────────────────────────────────────────
+const WAITING_PATTERNS = [
+  /\[y\/N\]/i, /\[Y\/n\]/i, /\(y\/n\)/i,
+  /Press Enter/i,
+  /Continue\?/i,
+  /Do you want to/i,
+  /Would you like/i,
+  /Proceed\?/i,
+  /\? .{1,60}[›>❯]\s*$/m,  // inquirer / Claude Code prompts
+  // Claude Code 承認待ちパターン
+  /Yes,?\s+allow/i,
+  /No,?\s+don['']t allow/i,
+  /Allow\s+(once|always|this)/i,
+  /\bAllow\b.{0,40}\?/i,
+  /Deny\b/i,
+  /Yes\s*\/\s*No/i,
+  /❯\s*(Yes|No|Allow|Deny)/,
+  /›\s*(Yes|No|Allow|Deny)/,
+  /\[\s*A\s*\]llow/i,
+  /\[\s*D\s*\]eny/i,
+  /approve.*\(y\/n\)/i,
+  // NOTE: /permission/i は削除 — Claude Code の UI フッター "bypass permissions on" に誤反応するため
+  // 日本語の確認待ちパターン（vk-kore など、Claude が確認を求めて中断する場面で出る文言）
+  /ご確認(?:を|ください|お願い)/,
+  /続行しますか/,
+  /進めて(?:よろしい|よい)/,
+  /(?:よろしい|いかが)(?:でしょうか|ですか)[。？?]?\s*$/m,
+  // マージ待ちパターン（vk-kore の PR 作成後・マージ判断委譲のタイミング）
+  /マージ(?:判断|してください|してもよろしい)/,
+  /マージ.{0,30}(?:ご判断|お願い|よろしい|お任せ)/,
+  // recap / 追加の確認待ち文言（issue #32）。
+  // Claude Code が "※ recap: …承認待ちです。…(disable recaps in /config)" のような
+  // 振り返りメッセージを最後に挟むケースで、本文末尾が "承認待ち" や "委任します"
+  // のような形になる。これらの「次アクションをユーザーに委ねている」言い回しを拾う。
+  /(?:承認|回答|ご判断|ご返答|お返事|ご指示|ご連絡)(?:を)?(?:お)?待ち/,
+  /(?:いただけ|いただい)たら.{0,30}(?:委任|お願い|進め|実装)/,
+  /(?:お任せ|ご判断)(?:します|ください|いただけ)/,
+  /(?:お待ち|待って)(?:しています|います|ます)/,
+  // AskUserQuestion / 数字選択肢の UI 検知（issue #46）。
+  // Claude Code の AskUserQuestion は「❯ 1. … / 2. …」の選択肢と
+  // 「Enter to select / ↑/↓ to navigate / Esc to cancel」のフッターが固定で出る。
+  // 既存パターンは ASCII `?` と `❯ Yes|No|Allow|Deny` しか拾えず取りこぼしていた。
+  /Enter\s+to\s+select/i,
+  /[↑↓]\/[↑↓]\s+to\s+navigate/,
+  /Esc\s+to\s+cancel/i,
+  /❯\s*\d+\.\s/,  // `❯ 1. ラベル` 形式（任意ラベルの数字選択肢）
+  // 全角「？」で終わる質問文。AskUserQuestion 以外の TUI / 日本語プロンプト
+  // でも全角？で末尾するケースを拾うための補助パターン。
+  // Claude Code の AskUserQuestion 自体は上の `Enter to select` フッターで
+  // 確定検知できるため、ここは網羅性ではなく **誤検知抑制** を優先して
+  // `m` フラグ無しでバッファ全体の末尾にのみアンカーする。
+  // `m` を付けると `lastLines` バッファ（最大 80 行）に残る過去の質問行に
+  // 反応して running 中も waiting に張り付くため、その挙動を避ける。
+  /[？]\s*$/,
+];
+
+function matchesWaiting(cleanBuffer) {
+  return WAITING_PATTERNS.some(p => p.test(cleanBuffer));
+}
+
+function nextWaitingState({ prev, matches }) {
+  // 出力再評価では waiting を解除しない。解除はユーザー入力時の明示クリアだけに集約する。
+  return matches || prev;
+}
+
+module.exports = {
+  WAITING_PATTERNS,
+  matchesWaiting,
+  nextWaitingState,
+};
